- 전체 Model domain 레이어의 model 패키지 아래로 이동 - data 레이어의 repository 파일들에서 interface 삭제 후 명칭 repositoryImpl 형식으로 변경
