+++ app/src/main/java/com/gyleedev/chatchat/domain/model/ChatRoomDataWithAllRelatedUsersAndMessage.kt
@@ -1,4 +1,4 @@
-package com.gyleedev.chatchat.domain
+package com.gyleedev.chatchat.domain.model
 
 data class ChatRoomDataWithAllRelatedUsersAndMessage(
     val chatRoomLocalData: ChatRoomLocalData,
