+++ app/src/main/java/com/gyleedev/chatchat/domain/model/ChatRoomLocalDataWrapper.kt
@@ -1,4 +1,4 @@
-package com.gyleedev.chatchat.domain
+package com.gyleedev.chatchat.domain.model
 
 sealed interface ChatRoomLocalDataWrapper {
     data object Failure : ChatRoomLocalDataWrapper
