+++ app/src/main/java/com/gyleedev/chatchat/domain/model/LogInState.kt
@@ -1,4 +1,4 @@
-package com.gyleedev.chatchat.domain
+package com.gyleedev.chatchat.domain.model
 
 sealed interface LogInState {
     data class Success(
