+++ app/src/main/java/com/gyleedev/chatchat/domain/model/MessageRemoteData.kt
@@ -1,7 +1,6 @@
-package com.gyleedev.chatchat.data.model
+package com.gyleedev.chatchat.domain.model
 
 import com.google.gson.annotations.SerializedName
-import com.gyleedev.chatchat.domain.MessageType
 
 // 기본값 설정 안해주면 crash남
 data class MessageRemoteData(
