+++ app/src/main/java/com/gyleedev/chatchat/domain/model/RelatedUserLocalData.kt
@@ -1,29 +1,27 @@
-package com.gyleedev.chatchat.data.model
+package com.gyleedev.chatchat.domain.model
 
 import com.google.gson.annotations.SerializedName
-import com.gyleedev.chatchat.domain.RelatedUserLocalData
-import com.gyleedev.chatchat.domain.UserRelationState
 
-// Relations 정보
 // 기본값 설정 안해주면 crash남
-data class RelatedUserRemoteData(
+data class RelatedUserLocalData(
+    @SerializedName("id") val id: Long = 0,
     @SerializedName("email") val email: String = "",
     @SerializedName("name") val name: String = "",
     @SerializedName("uid") val uid: String = "",
     @SerializedName("picture") val picture: String = "",
     @SerializedName("status") val status: String = "",
     @SerializedName("userRelation") val userRelation: UserRelationState = UserRelationState.UNKNOWN,
-    @SerializedName("favoriteState") val favoriteState: Boolean = false
+    @SerializedName("favoriteState") val favoriteState: Boolean = false,
+    @SerializedName("favoriteNumber") val favoriteNumber: Long? = null
 )
 
-fun RelatedUserRemoteData.toRelatedUserLocalData(): RelatedUserLocalData {
-    return RelatedUserLocalData(
-        id = 0L,
+fun RelatedUserLocalData.toRemoteData(): RelatedUserRemoteData {
+    return RelatedUserRemoteData(
         email = email,
-        name = name,
         uid = uid,
-        picture = picture,
         status = status,
+        name = name,
+        picture = picture,
         userRelation = userRelation,
         favoriteState = favoriteState
     )
