+++ app/src/main/java/com/gyleedev/chatchat/domain/model/RelatedUserRemoteData.kt
@@ -1,28 +1,27 @@
-package com.gyleedev.chatchat.domain
+package com.gyleedev.chatchat.domain.model
 
 import com.google.gson.annotations.SerializedName
-import com.gyleedev.chatchat.data.model.RelatedUserRemoteData
 
+// Relations 정보
 // 기본값 설정 안해주면 crash남
-data class RelatedUserLocalData(
-    @SerializedName("id") val id: Long = 0,
+data class RelatedUserRemoteData(
     @SerializedName("email") val email: String = "",
     @SerializedName("name") val name: String = "",
     @SerializedName("uid") val uid: String = "",
     @SerializedName("picture") val picture: String = "",
     @SerializedName("status") val status: String = "",
     @SerializedName("userRelation") val userRelation: UserRelationState = UserRelationState.UNKNOWN,
-    @SerializedName("favoriteState") val favoriteState: Boolean = false,
-    @SerializedName("favoriteNumber") val favoriteNumber: Long? = null
+    @SerializedName("favoriteState") val favoriteState: Boolean = false
 )
 
-fun RelatedUserLocalData.toRemoteData(): RelatedUserRemoteData {
-    return RelatedUserRemoteData(
+fun RelatedUserRemoteData.toRelatedUserLocalData(): RelatedUserLocalData {
+    return RelatedUserLocalData(
+        id = 0L,
         email = email,
-        uid = uid,
-        status = status,
         name = name,
+        uid = uid,
         picture = picture,
+        status = status,
         userRelation = userRelation,
         favoriteState = favoriteState
     )
