+++ app/src/main/java/com/gyleedev/chatchat/domain/model/SearchUserResult.kt
@@ -1,4 +1,4 @@
-package com.gyleedev.chatchat.domain
+package com.gyleedev.chatchat.domain.model
 
 sealed interface SearchUserResult {
     data class Success(
