+++ app/src/main/java/com/gyleedev/chatchat/domain/model/UserChatRoomData.kt
@@ -1,4 +1,4 @@
-package com.gyleedev.chatchat.domain
+package com.gyleedev.chatchat.domain.model
 
 import com.google.gson.annotations.SerializedName
